homework/requirements.txt
asgiref==3.5.2
Django==3.2
django-cleanup==6.0.0
python-decouple==3.6
pytz==2022.4
sorl-thumbnail==12.9.0
sqlparse==0.4.3
